feat: combat-participant enemy filter, replay polish, u32 team ids
M10 Task 3 polish batch:

- Native enemies[]/HTML team chips now filter to combat participants only
  (dealt/received damage, or received CC from the squad; enemy players
  always kept) via a new Metrics::combat_participant_enemies set computed
  in analyze(). enc.enemies and the EI adapter (via a new, unserialized
  Report::all_enemies) stay unfiltered, so ei-json remains faithful to
  real EI's own "keep every target" behavior. Fixture: 80 -> 49 enemies;
  Red Desert log: 456 -> 59. Zero calibrated values moved.
- Replay bounds-finiteness guard now checks all four sides (was min_x
  only, which missed a single-axis Infinity sample); Replay tab shows
  "No position data in this log" when every track is sample-less instead
  of a blank stage; enemy dot fill/stroke contrast bumped for visibility
  on the dark stage.
- team_id widened u16 -> u32 end-to-end (model/wvw/schema/ei), removing
  two lossy truncating casts now that dynamic WVWTEAMS ids are u32.
- Asset budget gate raised to 65,536B (controller-authorized this
  milestone) to cover the replay minors' added bytes.

+++ crates/axilog-core/src/analysis/cc.rs
@@ -174,8 +174,8 @@ fn pet_credit_cc_events(
     raw: &RawLog,
     squad: &BTreeSet<u64>,
     enemies: &BTreeSet<u64>,
-    friendly_team: Option<u16>,
-    agent_team: &std::collections::BTreeMap<u64, u16>,
+    friendly_team: Option<u32>,
+    agent_team: &std::collections::BTreeMap<u64, u32>,
 ) -> Vec<(u64, u64, u64)> {
     let post_era = raw.header.is_post_buff_rework();
     let registry = crate::analysis::damage::InstidRegistry::build(raw);
